Organize ingredients into subfolders
