+++ src/main/java/mezz/jei/plugins/vanilla/VanillaPlugin.java
@@ -31,12 +31,12 @@ import mezz.jei.plugins.vanilla.furnace.FuelRecipeMaker;
 import mezz.jei.plugins.vanilla.furnace.FurnaceFuelCategory;
 import mezz.jei.plugins.vanilla.furnace.FurnaceSmeltingCategory;
 import mezz.jei.plugins.vanilla.furnace.SmeltingRecipeMaker;
-import mezz.jei.plugins.vanilla.ingredients.FluidStackHelper;
-import mezz.jei.plugins.vanilla.ingredients.FluidStackListFactory;
-import mezz.jei.plugins.vanilla.ingredients.FluidStackRenderer;
-import mezz.jei.plugins.vanilla.ingredients.ItemStackHelper;
-import mezz.jei.plugins.vanilla.ingredients.ItemStackListFactory;
-import mezz.jei.plugins.vanilla.ingredients.ItemStackRenderer;
+import mezz.jei.plugins.vanilla.ingredients.fluid.FluidStackHelper;
+import mezz.jei.plugins.vanilla.ingredients.fluid.FluidStackListFactory;
+import mezz.jei.plugins.vanilla.ingredients.fluid.FluidStackRenderer;
+import mezz.jei.plugins.vanilla.ingredients.item.ItemStackHelper;
+import mezz.jei.plugins.vanilla.ingredients.item.ItemStackListFactory;
+import mezz.jei.plugins.vanilla.ingredients.item.ItemStackRenderer;
 import mezz.jei.startup.StackHelper;
 import mezz.jei.transfer.PlayerRecipeTransferHandler;
 import net.minecraft.client.gui.GuiRepair;
