+++ src/main/java/mezz/jei/plugins/vanilla/ingredients/item/ItemStackRenderer.java
@@ -1,4 +1,4 @@
-package mezz.jei.plugins.vanilla.ingredients;
+package mezz.jei.plugins.vanilla.ingredients.item;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
