+++ src/main/java/mezz/jei/plugins/vanilla/ingredients/item/package-info.java
@@ -1,0 +1,9 @@
+@ParametersAreNonnullByDefault
+@FieldsAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+package mezz.jei.plugins.vanilla.ingredients.item;
+
+import mcp.MethodsReturnNonnullByDefault;
+import mezz.jei.util.FieldsAreNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
